requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
cosine_annealing_warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
filelock==3.13.1
fsspec==2023.12.2
idna==3.6
iterative-stratification==0.1.7
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
pandas==2.2.0
pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
